bigint numbers in database

+++ app/src/main/java/com/example/numberproject/data/Number.kt
@@ -2,11 +2,13 @@ package com.example.numberproject.data
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import androidx.room.TypeConverter
+import java.math.BigInteger
 
 @Entity(tableName = "number")
 data class Number (
     @PrimaryKey(autoGenerate = true)
     val id : Int = 0,
-    val number : Int,
+    val number : BigInteger,
     val fact : String?,
-    )
+)
